+++ infra/redis/src/main/kotlin/kr/pe/hws/stockmanager/redis/hash/RestKisToken.kt
@@ -1,6 +1,5 @@
 package kr.pe.hws.stockmanager.redis.hash
 
-import kr.pe.hws.stockmanager.domain.token.ApiToken
 import org.springframework.data.annotation.Id
 import org.springframework.data.redis.core.RedisHash
 import org.springframework.data.redis.core.TimeToLive
@@ -13,14 +12,4 @@ class RestKisToken(
     val token_type: String,
     @TimeToLive
     val expires_in: Int,
-) {
-    // 기본 생성자 추가
-    constructor() : this("", "", "", 0)
-
-    constructor(token: ApiToken.KisToken) : this(
-        access_token = token.accessToken,
-        access_token_token_expired = token.accessTokenTokenExpired,
-        token_type = token.tokenType,
-        expires_in = (token.expiresIn -10000)
-    )
-}
+)
